CryptoProject/requirements.txt
streamlit
argon2-cffi
cryptography
ascon
numpy
opencv-python
Pillow